feat(features/cards): popup to modal
