+++ src/ui/organisms/index.js
@@ -1,4 +1,4 @@
 export { PrimitiveFooter } from "./primitive-footer"
 export { ItemsList } from "./items-list"
 export { ConditionalList } from "./conditional-list"
-export { PopUp } from "./popup"
+export { Modal } from "./modal"
